comments from arvind
